requirements.txt
browsermob_proxy==0.8.0
dnspython==2.7.0
func_timeout==4.3.5
numba==0.61.0
numpy==2.2.3
psutil==7.0.0
PyYAML==6.0.2
scapy==2.6.1
selenium==4.28.1
tqdm==4.66.5
transitions==0.9.2
